chore: the error is different now.
Data from the buffer is completely useless, I think I'll need to copy
the fd_write implementation and do the pointer dance.

Or use `fd_write` with a `fake` file descriptor to buffer data
between the guest and host (not sure what is the best approach)

Anyway, it is late already, time to stop!