chore: add a receive method to the exchange.
Instead of forcing all actors to work as P2P, the new API,
allows for simpler clients that connect to the stage network
via one exchange.

This, in turn, removes all the routing complexity from client
processess as they just need to call Deliver or Receive,
to interact with any node in the network.